+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add more apps for release notes.
+
+### Fixed
+
+- Fix fetching flatcar release notes.
+
 ## [4.21.0] - 2022-03-04
 
 ### Changed